generalize help parser; add help messages to books
